Split out measurements/api
